Pull helper function into words util
